move Crypto to lib
